Support multiple repository projects
This allows to check ACL of projects containing multiple repositories with
the scm-api, for instance WorkflowJob using a global pipeline git repo
library and several code repositories will now be tested.
As long as one repository allows the permission, the access is granted
to the project (which might not be what everybody wants).